Replace ATtiny24 by AVR16EB14
While both controllers share the same footprint, the AVR16EB14
is quite a bit different. To begin, the footprints have VDD and
GND swapped. The AVRxEB uses UPDI for programming, so basically
only a single wire is needed. The GPIOs are scattered across
more ports. Finally, the entire clock system and ADC is
different.