Remove external temperature sensor
The external sensor was initially designed as an option to allow for a
more accurate temperature sensor. To save part costs, it has never
actually been implemented.

With the AVRxEB, the internal temperature sensor comes pre-calibrated
(through fuse values), so it is safe to get rid of that idea
entirely.